pyproject.toml
[tool.poetry]
name = "credit-card-frauds"
version = "0.1.0"
description = "Using Machine Learning to predict Credit Card Frauds."
authors = ["Roberto Fierimonte <roberto.fierimonte@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
numpy = "==1.23"
numpy-financial = "==1.0.0"
pandas = "==1.4.3"
scikit-learn = "==1.1.2"
xgboost = "==1.6.1"
probatus = "==1.8.9"
imbalanced-learn = "==0.9.1"
optuna = "==3.0.3"
pandera = "~=0.6.5"
pydantic = "~=1.9.0"
Jinja2 = ">3"
holidays = "^0.25"
loguru = "^0.7.0"
lightgbm = "^3.3.5"
kfp = "^1.8.22"
google-cloud-pipeline-components = "^1.0.43"
google-cloud-aiplatform = "^1.25.0"
google-cloud-storage = "^2.9.0"
google-cloud-bigquery = {version = "^3.10.0", extras = ["pandas"]}
google-cloud-logging = "^3.5.0"
pyarrow = "^12.0.0"
fsspec = "^2023.6.0"
gcsfs = "^2023.6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
commitizen = "^2.37.0"
black = {version = "^22.10.0", allow-prereleases = true}
pytest = "^7.3.1"
opendatasets = "^0.1.22"
requests = "<2.31.0"
matplotlib = "==3.5.1"
seaborn = "==0.12.1"
google-cloud-bigquery-storage = "^2.20.0"

[tool.poetry.group.test.dependencies]
ipykernel = "^6.17.0"
xmlrunner = "^1.7.7"
unittest-xml-reporting = "^3.2.0"

[tool.poetry.group.cicd.dependencies]
pre-commit = "^2.20.0"
loguru = "^0.7.0"
gitpython = "^3.1.31"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
